small bug fixes when under heavy multithreading

+++ Scripts/convert_sound_zappa.py
@@ -2,6 +2,7 @@
 
 import os
 import util
+import time
 
 def GetVGAudioCli():
 	utilities_path = util.GetUtilitiesPath()
@@ -19,37 +20,40 @@ def XWM2WAV(filename_xwm, filename_wav):
 	util.RunCommandLine(commandLine)
 	
 def WAV2DSP(filename_wav, filename_dsp0, filename_dsp1):
-	VGAudioCli = GetVGAudioCli()
-	commandLine = [VGAudioCli, "-i:0", filename_wav, filename_dsp0]
-	util.RunCommandLine(commandLine)
-
 	with open(filename_wav, "rb") as wav_file:
 		wav_file.seek(0x16)
 		channel_count = int.from_bytes(wav_file.read(1), byteorder = 'little', signed = False)
-		if channel_count > 1:
-			commandLine = [VGAudioCli, "-i:1", filename_wav, filename_dsp1]
-			util.RunCommandLine(commandLine)
+		
+	VGAudioCli = GetVGAudioCli()
+	commandLine = [VGAudioCli, "-i:0", filename_wav, filename_dsp0]
+	util.RunCommandLine(commandLine)
+	if channel_count > 1:
+		commandLine = [VGAudioCli, "-i:1", filename_wav, filename_dsp1]
+		util.RunCommandLine(commandLine)
+		channel_count = 2
+	
+	return channel_count
 
-def ConvertDSP(dsp_data):
-	dsp_data[0x00:0x19:4], dsp_data[0x01:0x1A:4], dsp_data[0x02:0x1B:4], dsp_data[0x03:0x1C:4] = dsp_data[0x03:0x1C:4], dsp_data[0x02:0x1B:4], dsp_data[0x01:0x1A:4], dsp_data[0x00:0x19:4]
-	dsp_data[0x1C:0x5F:2], dsp_data[0x1D:0x60:2] = dsp_data[0x1D:0x60:2], dsp_data[0x1C:0x5F:2]
+def ConvertDSP(dsp_data, base):
+	dsp_data[base+0x00:base+0x19:4], dsp_data[base+0x01:base+0x1A:4], dsp_data[base+0x02:base+0x1B:4], dsp_data[base+0x03:base+0x1C:4] = \
+	dsp_data[base+0x03:base+0x1C:4], dsp_data[base+0x02:base+0x1B:4], dsp_data[base+0x01:base+0x1A:4], dsp_data[base+0x00:base+0x19:4]
+	dsp_data[base+0x1C:base+0x5F:2], dsp_data[base+0x1D:base+0x60:2] = \
+	dsp_data[base+0x1D:base+0x60:2], dsp_data[base+0x1C:base+0x5F:2]
 
-def DSP2MCADPCM(filename_dsp0, filename_dsp1, sound_file):
+def DSP2MCADPCM(filename_dsp0, filename_dsp1, channels, sound_file):
 	header_single = b'\x01\x00\x00\x00\x0C\x00\x00\x00'
 	header_stereo = b'\x02\x00\x00\x00\x14\x00\x00\x00'
 	
-	with open(filename_dsp0, "rb") as dsp0_file:
-		dsp0_data = bytearray(dsp0_file.read())
-		dsp0_size = len(dsp0_data)
-		ConvertDSP(dsp0_data)
-		
-	if os.path.exists(filename_dsp1):
-		with open(filename_dsp1, "rb") as dsp1_file:
-			dsp1_data = bytearray(dsp1_file.read())
-			dsp1_size = len(dsp1_data)
-			ConvertDSP(dsp1_data)
-		
+	with open(filename_dsp0, "rb") as dsp0_file: dsp0_data = bytearray(dsp0_file.read())
+	dsp0_size = len(dsp0_data)
+	ConvertDSP(dsp0_data, 0x00)
+
+	if channels == 2:
+		with open(filename_dsp1, "rb") as dsp1_file: dsp1_data = bytearray(dsp1_file.read())
+		dsp1_size = len(dsp1_data)
 		dsp1_offset = 0x14 + dsp0_size
+		ConvertDSP(dsp1_data, 0x00)
+
 		sound_file.write(header_stereo)
 		sound_file.write(dsp0_size.to_bytes(4, byteorder = 'little', signed = False))
 		sound_file.write(dsp1_offset.to_bytes(4, byteorder = 'little', signed = False))
@@ -64,9 +68,8 @@ def DSP2MCADPCM(filename_dsp0, filename_dsp1, sound_file):
 
 def ConvertSound_Internal(filepath_without_extension):
 	filename_mcadpcm = filepath_without_extension + ".mcadpcm"
-	filename_dsp0 = filepath_without_extension + "__CH0__.dsp"
-	filename_dsp1 = filepath_without_extension + "__CH1__.dsp"
-
+	filename_dsp0 = filepath_without_extension + "_CH0_.dsp"
+	filename_dsp1 = filepath_without_extension + "_CH1_.dsp"	
 	filename_wav = filepath_without_extension + ".wav"
 	filename_xwm = filepath_without_extension + ".xwm"
 	filename_lip = filepath_without_extension + ".lip"
@@ -95,17 +98,18 @@ def ConvertSound_Internal(filepath_without_extension):
 			if lip_size == 0:
 				fuz_file.seek(0x08)
 				lip_size = int.from_bytes(fuz_file.read(0x04), byteorder = 'little', signed = False)
-				lip_data = fuz_file.read(lip_size) # if lip_size > 0
+				lip_data = fuz_file.read(lip_size)
 			else:
 				fuz_file.seek(0x12)				
 			if not (has_wav or has_xwm):
 				with open(filename_xwm, "wb") as xwm_file:
 					xwm_file.write(fuz_file.read())
+					has_xwm = True
 			
 	if not has_wav:
 		XWM2WAV(filename_xwm, filename_wav)
 
-	WAV2DSP(filename_wav, filename_dsp0, filename_dsp1)
+	channels = WAV2DSP(filename_wav, filename_dsp0, filename_dsp1)
 
 	# convert DSP to MCADPCM and save as MCADPCM or as FUZ if LIP exists
 	if lip_size > 0:
@@ -121,17 +125,17 @@ def ConvertSound_Internal(filepath_without_extension):
 			fuz_nx_file.write(voice_offset.to_bytes(4, byteorder = 'little', signed = False))
 			fuz_nx_file.write(lip_data)
 			fuz_nx_file.write(b'\x00' * lip_padding)
-			DSP2MCADPCM(filename_dsp0, filename_dsp1, fuz_nx_file)			
+			DSP2MCADPCM(filename_dsp0, filename_dsp1, channels, fuz_nx_file)			
 	else:
 		with open(filename_mcadpcm, "wb") as mcadpcm_file:
-			DSP2MCADPCM(filename_dsp0, filename_dsp1, mcadpcm_file)
+			DSP2MCADPCM(filename_dsp0, filename_dsp1, channels, mcadpcm_file)
 
 	# clean up temporary files	
-	util.RemoveFile(filename_dsp0)
-	util.RemoveFile(filename_dsp1)
 	util.RemoveFile(filename_wav) 
-	util.RemoveFile(filename_xwm)
-	util.RemoveFile(filename_lip)
+	util.RemoveFile(filename_dsp0)
+	if channels == 2: util.RemoveFile(filename_dsp1)
+	if has_xwm: util.RemoveFile(filename_xwm)
+	if has_lip: util.RemoveFile(filename_lip)
 	if lip_size == 0: util.RemoveFile(filename_fuz)
 	
 	return True
